+++ tests/testthat/test-generateCohortSet.R
@@ -89,7 +89,11 @@ test_that("duckdb cohort generation with attrition", {
   # test readCohortSet with non existing directory
   expect_error(readCohortSet(system.file("cohorts99", package = "CDMConnector", mustWork = FALSE)))
 
-  cdm <- generateCohortSet(cdm, cohortSet, name = "chrt0", computeAttrition = TRUE)
+  cdm <- generateCohortSet(cdm,
+                           cohortSet,
+                           name = "chrt0",
+                           computeAttrition = TRUE,
+                           overwrite = TRUE)
 
   expect_true("chrt0" %in% listTables(con, schema = write_schema))
 
@@ -163,7 +167,11 @@ test_that("SQL Server cohort generation with attrition", {
   cohortSet <- cohortSet[3,]
   expect_s3_class(cohortSet, "CohortSet")
 
-  cdm <- generateCohortSet(cdm, cohortSet, name = "chrt0", computeAttrition = TRUE)
+  cdm <- generateCohortSet(cdm,
+                           cohortSet,
+                           name = "chrt0",
+                           computeAttrition = TRUE,
+                           overwrite = TRUE)
 
   expect_true("chrt0" %in% listTables(con, schema = write_schema))
 
@@ -220,7 +228,11 @@ test_that("Redshift cohort generation with attrition", {
   cohortSet <- cohortSet[3,]
   expect_s3_class(cohortSet, "CohortSet")
 
-  cdm <- generateCohortSet(cdm, cohortSet, name = "chrt0", computeAttrition = TRUE)
+  cdm <- generateCohortSet(cdm,
+                           cohortSet,
+                           name = "chrt0",
+                           computeAttrition = TRUE,
+                           overwrite = TRUE)
 
   expect_true("chrt0" %in% listTables(con, schema = write_schema))
 
@@ -274,7 +286,11 @@ test_that("Postgres cohort generation with attrition", {
   cohortSet <- cohortSet[3,]
   expect_s3_class(cohortSet, "CohortSet")
 
-  cdm <- generateCohortSet(cdm, cohortSet, name = "chrt0", computeAttrition = TRUE)
+  cdm <- generateCohortSet(cdm,
+                           cohortSet,
+                           name = "chrt0",
+                           computeAttrition = TRUE,
+                           overwrite = TRUE)
 
   expect_true("chrt0" %in% listTables(con, schema = write_schema))
 
@@ -305,37 +321,61 @@ test_that("Postgres cohort generation with attrition", {
 })
 
 
-# test_that("cohort generation works on spark", {
-#
-#   skip_if_not("Databricks" %in% odbc::odbcListDataSources()$name)
-#   skip("manual test")
-#   skip_if_not_installed("CirceR")
-#
-#   con <- DBI::dbConnect(odbc::odbc(), dsn = "Databricks")
-#
-#   write_schema <- "omop531results"
-#   cdm_schema <- "omop531"
-#
-#   cdm <- cdmFromCon(con,
-#                     cdmSchema = cdm_schema,
-#                     cdmTables = tbl_group("default"),
-#                     writeSchema = write_schema)
-#
-#   cohortSet <- readCohortSet(system.file("cohorts2", package = "CDMConnector", mustWork = TRUE))
-#   expect_equal(nrow(cohortSet), 2)
-#
-#   cdm <- addCohortTable(cdm, name = "cohorts", overwrite = TRUE)
-#   expect_true("cohorts" %in% names(cdm))
-#
-#   cdm <- generateCohortSet(cdm, cohortSet, cohortTableName = "cohorts", overwrite = TRUE)
-#   df <- cdm$cohorts %>% head() %>% dplyr::collect()
-#   expect_s3_class(df, "data.frame")
-#   expect_true(all(names(df) == c("cohort_definition_id", "subject_id", "cohort_start_date", "cohort_end_date")))
-#
-#   DBI::dbRemoveTable(con, DBI::Id(schema = write_schema, table = "cohorts"))
-#   expect_false("cohorts" %in% listTables(con, schema = write_schema))
-#   DBI::dbDisconnect(con)
-# })
+test_that("cohort generation works on spark", {
+
+  skip_if_not("Databricks" %in% odbc::odbcListDataSources()$name)
+  skip("manual test")
+  skip_if_not_installed("CirceR")
+
+  con <- DBI::dbConnect(odbc::odbc(), dsn = "Databricks")
+
+  write_schema <- "omop531results"
+  cdm_schema <- "omop531"
+
+  cdm <- cdmFromCon(con,
+                    cdmSchema = cdm_schema,
+                    cdmTables = tbl_group("default"),
+                    writeSchema = write_schema)
+
+  cohortSet <- readCohortSet(system.file("cohorts2", package = "CDMConnector", mustWork = TRUE))
+  expect_equal(nrow(cohortSet), 3)
+  cohortSet <- cohortSet[3,]
+  expect_s3_class(cohortSet, "CohortSet")
+  # debugonce(generateCohortSet)
+  options(sqlRenderTempEmulationSchema = write_schema)
+  cdm <- generateCohortSet(cdm,
+                           cohortSet,
+                           name = "chrt0",
+                           computeAttrition = TRUE,
+                           overwrite = TRUE)
+
+  expect_true("chrt0" %in% listTables(con, schema = write_schema))
+
+  expect_true("GeneratedCohortSet" %in% class(cdm$chrt0))
+  df <- cdm$chrt0 %>% head() %>% dplyr::collect()
+  expect_s3_class(df, "data.frame")
+  expect_true(nrow(df) > 0)
+  expect_true(all(c("cohort_definition_id", "subject_id", "cohort_start_date", "cohort_end_date") %in% names(df)))
+
+  attrition_df <- dplyr::collect(cohortAttrition(cdm$chrt0))
+  expect_s3_class(attrition_df, "data.frame")
+  expect_true(nrow(attrition_df) > 0)
+  expect_true("excluded_records" %in% names(attrition_df))
+  expect_s3_class(dplyr::collect(cohortSet(cdm$chrt0)), "data.frame")
+  counts <- dplyr::collect(cohortCount(cdm$chrt0))
+  expect_s3_class(counts, "data.frame")
+  expect_equal(nrow(counts), nrow(cohortSet))
+
+  # drop tables
+  DBI::dbRemoveTable(con, DBI::Id(schema = write_schema, table = "chrt0"))
+  expect_false("chrt0" %in% listTables(con, schema = write_schema))
+
+  DBI::dbRemoveTable(con, DBI::Id(schema = write_schema, table = "chrt0_count"))
+  DBI::dbRemoveTable(con, DBI::Id(schema = write_schema, table = "chrt0_set"))
+  DBI::dbRemoveTable(con, DBI::Id(schema = write_schema, table = "chrt0_attrition"))
+
+  DBI::dbDisconnect(con)
+})
 
 
 # Test readCohortSet ----
@@ -345,3 +385,36 @@ test_that("ReadCohortSet gives informative error when pointed to a file", {
   expect_error(readCohortSet(path), "not a directory")
 })
 
+test_that("Generation from Capr Cohorts", {
+  skip_if_not(eunomia_is_available())
+  skip_if_not_installed("Capr")
+  library(Capr)
+
+  con <- DBI::dbConnect(duckdb::duckdb(), dbdir = eunomia_dir())
+  cdm <- CDMConnector::cdm_from_con(
+    con = con,
+    cdm_schema = "main",
+    write_schema = "main"
+  )
+
+  gibleed_cohort_definition <- cohort(
+    entry = condition(cs(descendants(192671))),
+    attrition = attrition(
+      "no RA" = withAll(
+        exactly(0,
+                condition(cs(descendants(80809))),
+                duringInterval(eventStarts(-Inf, Inf))))
+    )
+  )
+
+  cdm <- generateCohortSet(
+    cdm,
+    list(gibleed = gibleed_cohort_definition),
+    name = "gibleed",
+    computeAttrition = TRUE,
+    overwrite = TRUE
+  )
+
+  expect_gt(nrow(dplyr::collect(cdm$gibleed)), 10)
+  DBI::dbDisconnect(con, shutdown = TRUE)
+})